Add git-archive --remote= support to gitosis-serve.
Now read access implies ability to grab archives remotely.
To allow that through git-daemon (for symmetricality),
run "git config --system daemon.uploadarch true".
Note that as of this writing, git documentation lies about
the name of the variable.

NOTE: This does not work and will not work until git-upload-archive
support is added to git-shell. Once that is done, this can be merged.